requirements.txt
livekit-agents[openai, cartesia, elevenlabs, deepgram, silero, turn-detector,groq, mcp]>=1.0
python-dotenv>=1.0
duckduckgo-search>=8.0
livekit-api
fastapi[standard]
livekit